requirements.txt
locust
streamlit
pandas
matplotlib
gspread
oauth2client 
sarvamai
numpy
python-dotenv